requirements.txt
Django>=4
djangorestframework
gunicorn
whitenoise
psycopg2-binary